Improve error handling of UiText.kt, moved to Utils package in UI
