+++ app/src/main/java/com/realityexpander/tasky/presentation/register_screen/RegisterEvent.kt
@@ -1,7 +1,7 @@
 package com.realityexpander.tasky.presentation.register_screen
 
 import com.realityexpander.tasky.common.AuthToken
-import com.realityexpander.tasky.common.UiText
+import com.realityexpander.tasky.ui.util.UiText
 
 sealed class RegisterEvent {
     data class Loading(val isLoading: Boolean) : RegisterEvent()
